refactor: 점수 비교 로직 Participants로 이동
