+++ src/main/kotlin/blackjack/model/domain/participant/Player.kt
@@ -1,6 +1,5 @@
 package blackjack.model.domain.participant
 
-import blackjack.model.domain.GameResult
 import blackjack.model.domain.card.Card
 import blackjack.model.domain.card.Hand
 
@@ -13,27 +12,5 @@ data class Player(override val name: String) : Participants() {
 
     override fun showStartCards(): List<Card> {
         return hand.cards
-    }
-
-    fun compareScores(otherHand: Hand): GameResult =
-        when {
-            otherHand.isBlackJack() && hand.isBlackJack() -> GameResult.Draw
-            hand.isBlackJack() -> GameResult.BlackjackWin
-            hand.isBust() -> GameResult.Lose
-            otherHand.isBust() -> GameResult.Win
-            otherHand.isBlackJack() -> GameResult.Lose
-            else -> compare(sumCardNumber, otherHand.getSumNumber())
-        }
-
-    private fun compare(
-        target: Int,
-        other: Int,
-    ): GameResult {
-        if (target < other) {
-            return GameResult.Lose
-        } else if (target > other) {
-            return GameResult.Win
-        }
-        return GameResult.Draw
     }
 }
